feat(playlist-maker): add new result node type
