+++ libs/shared/src/hooks/use-play-status.ts
@@ -1,19 +1,43 @@
 import { useEffect, useState } from 'react';
 
-export type PlayStatus = 'play' | 'pause';
+export enum PlayStatus {
+    Playing = 'playing',
+    Paused = 'paused',
+    Stopped = 'stopped',
+    Buffering = 'buffering',
+}
+
+const toPlayStatus = (state: Spicetify.PlayerState | undefined): PlayStatus => {
+    if (state === undefined) {
+        return PlayStatus.Stopped;
+    }
+
+    if (state.isPaused) {
+        return PlayStatus.Paused;
+    }
+
+    if (state.isBuffering) {
+        return PlayStatus.Buffering;
+    }
+
+    return PlayStatus.Playing;
+};
 
 export function usePlayStatus(): PlayStatus {
     const [playStatus, setPlayStatus] = useState<PlayStatus>(
-        Spicetify.Player.data.isPaused ? 'pause' : 'play',
+        toPlayStatus(Spicetify.Player.data),
     );
 
     useEffect(() => {
-        function handleStatusChange(event?: Event): void {
-            setPlayStatus(
-                (event as Event & { data: Spicetify.PlayerState }).data.isPaused
-                    ? 'pause'
-                    : 'play',
-            );
+        function handleStatusChange(
+            event?: Event & {
+                data: Spicetify.PlayerState;
+            },
+        ): void {
+            const playerState = event?.data;
+            const playStatus = toPlayStatus(playerState);
+
+            setPlayStatus(playStatus);
         }
 
         Spicetify.Player.addEventListener('onplaypause', handleStatusChange);
